+++ src/components/Typography/index.ts
@@ -1,0 +1,4 @@
+export { default as BodyText } from "./BodyText";
+export { default as H1 } from "./H1";
+export { default as H3 } from "./H3";
+export * from "./TypographyTypes";